fix CI: KeyPairGenerator import, Long cast, TouchStreamDecoder.MOVE_EMIT_PX qualification, file-level StreamHandle, ink consts to file scope, @file:OptIn(ExperimentalMaterial3Api), null-safe app list build

+++ app/src/main/java/com/jarvis/mobile/core/adb/SelfHostServer.kt
@@ -187,7 +187,7 @@ object SelfHostServer {
                 val n = stdout.read(buf)
                 if (n < 0) break
                 if (written + n > MAX_OUTPUT_BYTES) {
-                    written = MAX_OUTPUT_BYTES // stop streaming more, keep the exit path
+                    written = MAX_OUTPUT_BYTES.toLong() // stop streaming more, keep the exit path
                     break
                 }
                 out.writeInt(n)
